Mover wrapper_base to a dedicated file
