+++ src/wrapper/refs_wrapper.hpp
@@ -5,6 +5,7 @@
 #include <git2.h>
 
 #include "../wrapper/repository_wrapper.hpp"
+#include "../wrapper/wrapper_base.hpp"
 
 class reference_wrapper : public wrapper_base<git_reference>
 {
