+++ src/wrapper/wrapper_base.hpp
@@ -1,0 +1,35 @@
+#pragma once
+
+
+template <class T>
+class wrapper_base
+{
+public:
+    using resource_type = T;
+
+    wrapper_base(const wrapper_base&) = delete;
+    wrapper_base& operator=(const wrapper_base&) = delete;
+
+    wrapper_base(wrapper_base&& rhs)
+        : p_resource(rhs.p_resource)
+    {
+        rhs.p_resource = nullptr;
+    }
+    wrapper_base& operator=(wrapper_base&& rhs)
+    {
+        std::swap(p_resource, rhs.p_resource);
+        return *this;
+    }
+
+    operator resource_type*() const noexcept
+    {
+        return p_resource;
+    }
+
+protected:
+    // Allocation and deletion of p_resource must be handled by inheriting class.
+    wrapper_base() = default;
+    ~wrapper_base() = default;
+
+    resource_type* p_resource = nullptr;
+};
